Split CreamSymbolProcessor into per-annotation processors
CreamSymbolProcessor.kt had grown to 903 lines — far past the project's
500-line ceiling — by housing all seven annotation handlers plus helper
types and utilities in one class.

Move each processXxx into its own file under ksp/process/ as an
`internal fun CreamSymbolProcessor.processXxx(resolver)` extension, and
relocate the shared helpers (extractKDoc, createNewKotlinFile) to ksp/util/.
The annotation-local helper types (CopyMappingInfo, CombineMappingInfo,
CombineToSourceEntry, TargetSourcesMapForCombineTo) move next to the
processor that uses them. `options` / `codeGenerator` become `internal` so
the extensions can reach them.

Pure relocation — no behavior change. CreamSymbolProcessor.kt is now a
48-line dispatcher and every new file is within the line-count target. The
snapshot suite is byte-identical, confirming generated output is unchanged.
