+++ cream-ksp/src/main/kotlin/me/tbsten/cream/ksp/process/CopyToChildrenProcessor.kt
@@ -1,0 +1,118 @@
+package me.tbsten.cream.ksp.process
+
+import com.google.devtools.ksp.getAnnotationsByType
+import com.google.devtools.ksp.processing.Dependencies
+import com.google.devtools.ksp.processing.Resolver
+import com.google.devtools.ksp.symbol.ClassKind
+import com.google.devtools.ksp.symbol.KSAnnotated
+import com.google.devtools.ksp.symbol.KSAnnotation
+import com.google.devtools.ksp.symbol.KSClassDeclaration
+import com.google.devtools.ksp.symbol.KSDeclaration
+import com.google.devtools.ksp.symbol.KSType
+import com.google.devtools.ksp.validate
+import me.tbsten.cream.CopyToChildren
+import me.tbsten.cream.ksp.CreamSymbolProcessor
+import me.tbsten.cream.ksp.GenerateSourceAnnotation
+import me.tbsten.cream.ksp.InvalidCreamUsageException
+import me.tbsten.cream.ksp.transform.appendCombineToFunction
+import me.tbsten.cream.ksp.transform.appendCopyFunction
+import me.tbsten.cream.ksp.util.createNewKotlinFile
+import me.tbsten.cream.ksp.util.extractKDoc
+import me.tbsten.cream.ksp.util.fullName
+import me.tbsten.cream.ksp.util.isSealed
+import me.tbsten.cream.ksp.util.requireClassDeclaration
+import me.tbsten.cream.ksp.util.resolveToClassDeclaration
+import me.tbsten.cream.ksp.util.underPackageName
+
+internal fun CreamSymbolProcessor.processCopyToChildren(resolver: Resolver): List<KSAnnotated> {
+    val (copyToChildrenTargets, invalidCopyToChildrenTargets) =
+        resolver
+            .getSymbolsWithAnnotation(
+                annotationName = CopyToChildren::class.fullName,
+            ).partition { it.validate() }
+
+    copyToChildrenTargets.forEach { copyToChildren ->
+        val sourceSealedClass =
+            run {
+                if (copyToChildren !is KSClassDeclaration) {
+                    throw InvalidCreamUsageException(
+                        message =
+                            "@${CopyToChildren::class.simpleName} annotation must be applied to a sealed class/interface." +
+                                if (copyToChildren is KSDeclaration) {
+                                    copyToChildren.simpleName.asString() +
+                                        " is not sealed class/interface"
+                                } else {
+                                    ""
+                                },
+                        solution = (copyToChildren as? KSDeclaration)?.let { "Make ${it.fullName} a sealed class/interface." },
+                    )
+                }
+
+                if (!copyToChildren.isSealed()) {
+                    throw InvalidCreamUsageException(
+                        message = "@${CopyToChildren::class.simpleName} annotation must be applied to a sealed class/interface, but ${copyToChildren.isSealed()}",
+                        solution = "",
+                    )
+                }
+
+                copyToChildren
+            }
+        // Enclose notCopyToObject in runCatching because it may cause an error if notCopyToObject cannot be obtained.
+        val notCopyToObject =
+            runCatching {
+                val copyToChildrenAnnotation =
+                    sourceSealedClass.getAnnotationsByType(CopyToChildren::class).firstOrNull()
+                copyToChildrenAnnotation
+                    ?.notCopyToObject
+                    ?: options.notCopyToObject
+            }.getOrDefault(false)
+
+        val (kdocDescription, kdocExamples) =
+            sourceSealedClass.annotations
+                .firstOrNull {
+                    it.annotationType
+                        .resolve()
+                        .declaration.fullName == CopyToChildren::class.qualifiedName
+                }?.extractKDoc()
+                ?: ("" to emptyList())
+
+        val targetClasses = sourceSealedClass.getSealedSubclasses()
+
+        codeGenerator
+            .createNewKotlinFile(
+                dependencies =
+                    Dependencies(
+                        aggregating = true,
+                        sourceSealedClass.containingFile!!,
+                    ),
+                packageName = sourceSealedClass.packageName,
+                fileName = "CopyToChildren__${sourceSealedClass.underPackageName}",
+            ) {
+                it.appendLine("import me.tbsten.cream.*")
+                it.appendLine()
+
+                targetClasses.forEach { targetClass ->
+                    // generate sourceClass to sourceClass copy function
+                    it.appendCopyFunction(
+                        source = sourceSealedClass,
+                        target = targetClass,
+                        omitPackages =
+                            listOf(
+                                "kotlin",
+                                sourceSealedClass.packageName.asString(),
+                            ),
+                        options = options,
+                        generateSourceAnnotation =
+                            GenerateSourceAnnotation.CopyToChildren(
+                                annotationTarget = sourceSealedClass,
+                                kdocDescription = kdocDescription,
+                                kdocExamples = kdocExamples,
+                            ),
+                        notCopyToObject = notCopyToObject,
+                    )
+                }
+            }
+    }
+
+    return invalidCopyToChildrenTargets
+}
